// frontend/src/api/generated/models/PaymentHistoryWithPersonDTO.ts
export interface PaymentHistoryWithPersonDTO {
    /**
     * 
     * @type {number}
     * @memberof PaymentHistoryWithPersonDTO
     */
    id: number;
    /**
     * 
     * @type {number}
     * @memberof PaymentHistoryWithPersonDTO
     */
    contractId: number;
    /**
     * 
     * @type {PersonSmallDetailsDTO}
     * @memberof PaymentHistoryWithPersonDTO
     */
    person?: PersonSmallDetailsDTO | null;
    /**
     * 
     * @type {RoomWithApartmentDTO}
     * @memberof PaymentHistoryWithPersonDTO
     */
    room?: RoomWithApartmentDTO | null;
    /**
     * 
     * @type {string}
     * @memberof PaymentHistoryWithPersonDTO
     */
    payedDate?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PaymentHistoryWithPersonDTO
     */
    scopeDate?: string | null;
    /**
     * 
     * @type {number}
     * @memberof PaymentHistoryWithPersonDTO
     */
    amount: number;
    /**
     * 
     * @type {PaymentStatus}
     * @memberof PaymentHistoryWithPersonDTO
     */
    status: PaymentStatus;
}